Treat verification postconditions more like actions. In particular, allow declaring them symbolically with pathnames.
